+++ apps/api/src/ai/ai.module.ts
@@ -9,7 +9,6 @@ import { AiJudgeConfigurationValidatorService } from "src/ai/judge-configuration
 import { AiRepository } from "src/ai/repositories/ai.repository";
 import { RagRepository } from "src/ai/repositories/rag.repository";
 import { AiPracticeContentGeneratorService } from "src/ai/services/ai-practice-content-generator.service";
-import { AiPracticeJudgeConfigurationService } from "src/ai/services/ai-practice-judge-configuration.service";
 import { AiPracticeService } from "src/ai/services/ai-practice.service";
 import { AiRuntimeService } from "src/ai/services/ai-runtime.service";
 import { AiService } from "src/ai/services/ai.service";
@@ -32,7 +31,6 @@ import { StudentLessonProgressModule } from "src/studentLessonProgress/studentLe
     AiJudgeConfigurationGeneratorService,
     AiJudgeConfigurationGenerationWorkflowService,
     AiJudgeConfigurationValidatorService,
-    AiPracticeJudgeConfigurationService,
     AiPracticeContentGeneratorService,
     ChatService,
     AiRuntimeService,
